# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API simulador de escuela",
        "description": "Api donde podras ver, editar, crear y eliminar los datos de los estudiantes, profesores y materias, asi como tambien matricular un alumno a una materia.",
        "contact": {
            "email": "arturo@arturo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/docente": {
            "get": {
                "tags": [
                    "Docentes"
                ],
                "summary": "Listado de todos los registros de los Docentes",
                "operationId": "333fee026ff9e6356de198172c1cbd2c",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Juan"
                                                    },
                                                    "apellido": {
                                                        "type": "string",
                                                        "example": "Cruz"
                                                    },
                                                    "telefono": {
                                                        "type": "string",
                                                        "example": "998545213"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "juanC@gmail.com"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Docentes"
                ],
                "summary": "Registrar la información de un Docentes",
                "operationId": "6f96b78084ca67a74cdaf2f6463a5a06",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "apellido": {
                                                "type": "string"
                                            },
                                            "telefono": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "escribe el nombre",
                                    "apellido": "escribe el apellido",
                                    "telefono": "escribe el telefono",
                                    "email": "escribe el email"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Francisco"
                                        },
                                        "apellido": {
                                            "type": "string",
                                            "example": "Alcantara"
                                        },
                                        "telefono": {
                                            "type": "string",
                                            "example": "987654321"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "pachito@alcantara.xyz"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The apellidos field is required."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Objeto de errores"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/docente/{id}": {
            "get": {
                "tags": [
                    "Docentes"
                ],
                "summary": "Mostrar la información de un docente",
                "operationId": "37831e907cc4854634b06b90428f97bd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Francisco"
                                        },
                                        "apellido": {
                                            "type": "string",
                                            "example": "Alcantara"
                                        },
                                        "telefono": {
                                            "type": "string",
                                            "example": "987654321"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "pachito@alcantara.xyz"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Alumno] #id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Docentes"
                ],
                "summary": "Actualizar la información de un Docente",
                "operationId": "8f46da718f5a5078ca5d754deade6559",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "apellido": {
                                                "type": "string"
                                            },
                                            "telefono": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "escribe el nombre",
                                    "apellido": "escribe el apellido",
                                    "telefono": "escribe el telefono",
                                    "email": "escribe el email"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Francisco"
                                        },
                                        "apellido": {
                                            "type": "string",
                                            "example": "Alcantara"
                                        },
                                        "telefono": {
                                            "type": "string",
                                            "example": "987654321"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "pachito@alcantara.xyz"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The apellidos field is required."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Objeto de errores"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Docentes"
                ],
                "summary": "Eliminar la información de un Docente",
                "operationId": "bb047d7ca215ccf4f7c48cd3057f18b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "NO CONTENT (eliminado correctamente)"
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se pudo realizar correctamente la operación"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}